Some self-review.
